refactor: refactor intro login component
